Interpret "no timeout" as infinite. (#161)
The defaulting of our kingress timeouts was recently dropped: https://github.com/knative/networking/pull/22, though I have a PR to roll this back until it can follow proper process and get conformance coverage.

Currently net-contour implements this via Contour default, which uses Envoy defaults, which is 15s.

I believe the intended behavior of this is to be "no timeout", which is what this implements.