Add a scheme for warming Endpoints before rolling our kingress changes. (#166)
This is based on the sketch here: https://docs.google.com/document/d/11k0d0STquDFYBp_43I2AQQVDqQ22ZWfnon7RmJBcc2s/edit#

Contour has a race when it rolls out programming to Envoys where the routing
data can land prior to the endpoint data.  This is due to the fact that Contour
filters our Endpoints that aren't referenced by the routing layer to reduce the
amount of information it distributes, however, this means that when migrating
traffic to a new service that the Envoys have not seen before, the Envoys may
be directed to send traffic to Endpoints for which it has not yet received data.

This problem is especially bad for Knative because every Revision rolls out a
service that the Envoys haven't seen before, so virtually every rollout risks
a blip of 503s, which we actually see manifest fairly regularly in the e2e
testing, and which the TestUpdate kingress conformance test has exacerbated
for a while.

The solution in this pull request builds on the prior work to probe the Envoys
to ensure programming has been distributed.  It breaks the kingress
reconciliation down into two discrete phases:
1. Distribute artificial programming that references all of the Endpoints of
  our new programming.  Once probing has confirmed that to be Ready, then we...
2. Distribute the actual programming that references the new Endpoints, which
  we have confirmed the Envoys are aware of.

The implementation of this two-phased approach is done via a shallow recursion.
We create a secondary kingress targetting the appropriate Envoy pools with bogus
hostnames, and an annotation to block further recursion.  Once this "endpoint
probe" kingress has reported Ready, then we proceed with the remained of the
Contour reconciliation as we had it before, but now with the guarantee that the
Endpoints data is present.

Fixes: https://github.com/knative-sandbox/net-contour/issues/12

This should fix the longstanding TestUpdate issue, which I ran successfully (in
isolation) a half dozen times last night.

Related to: https://github.com/projectcontour/contour/issues/2151

I've tried to limit the scope of this change, but there are a handful of follow
ups I intend to make once this starts baking:

1. https://github.com/knative-sandbox/net-contour/issues/165: I noticed this
  making this change and it means that we are violating the "no calls on nops"
  rule for reconcilers.

2. Clean up the endpoints probe kingress resources.  I added a bool for this and
  a TODO, but I want to circle back to that once this has landed.  Technically
  what this PR does is sufficient for correctness, but it will scale better if
  we don't leave around the garbage programming.

3. Clean up the legacy Endpoints checking, which is subsumed by this and is now
  unnecessary complexity.