Fix a bad assumption in the HTTP Proxy logic. (#184)
The net-contour logic for translating kingress to httpproxy resources is a one-to-many
translation.  In fact, the arity of the http proxies can *change* dynamically over the
lifetime of a kingress.  So the logic to query, create, update, and delete httpproxy
resources is heavily label based.

Based on how Knative creates kingress resources, I do not believe it is possible to hit
this problem today, but given how we recursively use kingress for endpoint probing,
combined with how Knative uses kingress to invoke us, we hit this interesting corner case.

Essentially, the logic as it exists today assumes that the tuple `(kingress name, domain)`
is enough to uniquely identify the httpproxy; however, if a single kingress exposes the
same domain at multiple visibilities then this is insufficient.  We need the tuple
`(kingress name, domain, visibility)`.

Generally this wasn't a problem before because the external domain `foo.bar.example.com`
went to one visibility, while the other domain `foo.bar.svc.cluster.local` went to the other.
However, for the endpoint probe kingress, we create a host per service so that our probing
confirms that every endpoint is available.  Given that knative exposes the service by both
`foo.bar.example.com` AND `foo.bar.svc.cluster.local` this created a problem where we would
attempt to create the same endpoint probe httpproxy differing only in visibility.

There are two aspects to the fix for this:
1. Incorporate the visibility (class) into the name of the http proxy.
2. Label the http proxies with visibility, and filter on visibility when choosing which
  pre-existing httpproxy to update.

Despite changing resource names, this should be rollout safe because it only affects the
endpoint probe resource, which is new, and that resource doesn't exist in the steady state.

Fixes: https://github.com/knative-sandbox/net-contour/issues/176