Avoid thundering herd of probes during failover. (#231)
* Avoid thundering herd of probes during failover.

This change guards the status prober with a simple readiness check as an
optimization for the case of failing over and globally resyncing resources.
When a fresh controller pod slurps through ~100 kingress for the first time
we will currently probe ALL of them.  On a 10 Node cluster, with 1 public and
3 private hostnames, this works out to a lot of probes:

```
  100 * (1 * 10 + 3 * 10) = 4000
```

Factor in what we do with endpoint probing, and it adds roughly 2000 probes.

Currently, this is what the scale-100 test in serving is doing every 30s or so,
in addition to the standard work it is doing.

Fixes: https://github.com/knative-sandbox/net-contour/issues/203

Possibly related to: https://github.com/knative-sandbox/net-contour/issues/226

* With moar !Ready-ness