Ensure that endpoint probes avoid hostname collisions. (#247)
While investigating https://github.com/knative-sandbox/net-contour/issues/214 I noticed that the prober was seeing mismatched hashes when probing the hostnames of the endpoint probe kingress.  I might expect this if the kingress probe were updated, but I believe it was created fresh, so collisions are bizarre.  One plausible explanation of the 503s in the linked issue would be:

1. Main `kingress` reached steady state, endpoint probe `kingress` is deleted.
2. Main `kingress` is reprogrammed.
3. Fresh endpoint probe `kingress` is created with the relevant services.
4. HTTP Proxy from prior endpoint probe is leftover with the exitting service name, and the new endpoint probe updates it instead of creating a new HTTP Proxy.
5. Endpoint probe `kingress` successfully probes the updated HTTP Proxy that it does not own, and reports success.
6. Kubernetes GC catches up and deletes the HTTP Proxy owned by the prior endpoint probe.
7. Thinking the endpoints are adequately covered by the endpoint probe, the main kingress updates the main HTTP Proxy resources.
8. All references to the exitting service have been dropped from HTTP Proxy resources, so the cluster is removed from Envoy.
9. However, the "Cluster" corresponding to that service was still referenced by a "Route", so traffic hitting that "Route" returns a 503.

Reusing HTTP Proxy resources across generations of the kingress probe was never a requirement or goal, so this change simply avoids the possibility of collisions by incorporating the generation of the main `kingress` being probed into the hostnames using when constructing the endpoint probe.

Related: https://github.com/knative-sandbox/net-contour/issues/214