Disable Knative's Contour support for ExternalName K8s Services (#582)
* set the IP address on the KIngress load balancer status

* Disable HTTPProxy support for ExternalName K8s Services

ExternalName support is turned off by default in Contour given the advisory:
https://github.com/projectcontour/contour/security/advisories/GHSA-5ph6-qq5x-7jwc

The tl;dr use of ExternalName can expose sensitive services. Disabling this breaks DomainMapping
in Serving. As a workaround if a KIngress returns an LB IP Serving will manage this
endpoint using a headless service.

* Include a host rewrite test backed by a headless service

We'll want to move this to knative.dev/networking after the release

* exclude failing test in kind e2e

* quote test fixture names

* zap that error